[MODIF] header
If you enter no name for the project, the script will automatically take the name of the repo git or,
if the repo doesn't exist, the name of the folder where the file is located.
If you enter no name for the description, then the description will take the name of the file.